requirements.txt
djanjo
django-admin-argon-dashboard
google-play-scraper
appnope==0.1.0
backcall==0.1.0
certifi==2020.4.5.1
cffi==1.14.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
joblib==0.15.1
kiwisolver==1.2.0
matplotlib==3.2.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
nltk==3.5
numpy==1.18.1
pandas==1.0.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.4
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.5.14
six==1.14.0
torch==1.4.0
tqdm==4.46.0
traitlets==4.3.3
wcwidth==0.1.9